Move two context test methods into new test file
